+++ lib/prompt/io/choice.ex
@@ -3,63 +3,96 @@ defmodule Prompt.IO.Choice do
 
   alias __MODULE__
   alias IO.ANSI
-  import IO, only: [write: 1, read: 2]
+  import IO, only: [read: 2]
 
   @type t :: %Choice{
+          content: list(),
           default_answer: atom(),
           question: binary(),
           custom: any(),
           color: any(),
           background_color: any(),
-          trim: boolean()
+          trim: boolean(),
+          mask_line: boolean()
         }
 
   defstruct [
+    :content,
     :default_answer,
     :question,
     :custom,
     :color,
     :background_color,
-    :trim
+    :trim,
+    :mask_line
   ]
 
   @doc ""
   def new(question, custom, options) do
     [{k, _} | _rest] = custom
 
+    background_color =
+      options
+      |> Keyword.get(:background_color)
+      |> Prompt.IO.background_color()
+
     %Choice{
+      content: [:reset],
       color: Keyword.get(options, :color, IO.ANSI.default_color()),
+      background_color: background_color,
       trim: Keyword.get(options, :trim),
       default_answer: Keyword.get(options, :default_answer, k),
       question: question,
-      custom: custom
+      custom: custom,
+      mask_line: Keyword.get(options, :mask_line, false)
     }
   end
 
-  defimpl Prompt.IO do
+  def add_content(%Choice{content: content} = choice, to_add),
+    do: %{choice | content: content ++ [to_add]}
+
+  defimpl Prompt.IO.Terminal do
     def display(choice) do
       _ = Prompt.raw_mode_supported?() && :shell.start_interactive({:noshell, :cooked})
 
-      [
-        :reset,
-        background_color(choice),
-        choice.color,
-        "#{choice.question} #{choice_text(choice.custom, choice.default_answer)} ",
-        :reset,
-        without_newline(choice.trim)
-      ]
+      choice =
+        choice
+        |> Choice.add_content([
+          choice.background_color,
+          choice.color,
+          "#{choice.question} #{choice_text(choice.custom, choice.default_answer)} ",
+          :reset
+        ])
+        |> maybe_with_newline()
+
+      choice.content
       |> ANSI.format()
-      |> write()
+      |> Prompt.IO.write()
 
       choice
     end
 
     def evaluate(choice) do
       case read(:stdio, :line) do
-        :eof -> :error
-        {:error, _reason} -> :error
-        answer when is_binary(answer) -> _evaluate_choice(answer, choice)
-        answer when is_list(answer) -> _evaluate_choice(IO.chardata_to_string(answer), choice)
+        :eof ->
+          :error
+
+        {:error, _reason} ->
+          :error
+
+        answer when is_binary(answer) ->
+          if choice.mask_line do
+            Prompt.IO.Position.mask_line(1)
+          end
+
+          _evaluate_choice(answer, choice)
+
+        answer when is_list(answer) ->
+          if choice.mask_line do
+            Prompt.IO.Position.mask_line(1)
+          end
+
+          _evaluate_choice(IO.chardata_to_string(answer), choice)
       end
     end
 
@@ -92,14 +125,7 @@ defmodule Prompt.IO.Choice do
       end
     end
 
-    defp background_color(display) do
-      case display.background_color do
-        nil -> ANSI.default_background()
-        res -> String.to_atom("#{Atom.to_string(res)}_background")
-      end
-    end
-
-    defp without_newline(true), do: ""
-    defp without_newline(false), do: "\n"
+    defp maybe_with_newline(%Choice{trim: true} = choice), do: choice
+    defp maybe_with_newline(choice), do: Choice.add_content(choice, "\n")
   end
 end