+++ lib/prompt/io/display.ex
@@ -2,92 +2,121 @@ defmodule Prompt.IO.Display do
   @moduledoc false
 
   alias __MODULE__
-  import IO, only: [write: 1, read: 2]
+
+  import IO, only: [read: 2]
+  import Prompt.IO.ANSI
 
   @type t() :: %Display{
+          content: list(),
           text: binary(),
           color: Prompt.IO.Color.t(),
           background_color: Prompt.IO.Color.t(),
           trim: bool(),
           mask_line: bool(),
           from: atom(),
-          position: :left | :right
+          position: :left | :right,
+          alt_buffer: boolean()
         }
 
-  defstruct [:text, :color, :background_color, :trim, :mask_line, :from, :position]
+  defstruct [
+    :content,
+    :text,
+    :color,
+    :background_color,
+    :trim,
+    :mask_line,
+    :from,
+    :position,
+    :alt_buffer
+  ]
 
   def new(txt, options) do
+    background_color =
+      options
+      |> Keyword.get(:background_color)
+      |> Prompt.IO.background_color()
+
     %Display{
+      content: [:reset],
       text: txt,
       color: Keyword.get(options, :color, IO.ANSI.default_color()),
-      background_color: Keyword.get(options, :background_color),
+      background_color: background_color,
       mask_line: Keyword.get(options, :mask_line),
       trim: Keyword.get(options, :trim),
       from: Keyword.get(options, :from),
-      position: Keyword.get(options, :position)
+      position: Keyword.get(options, :position),
+      alt_buffer: Keyword.get(options, :alt_buffer, false)
     }
   end
 
-  defimpl Prompt.IO do
-    def display(text) do
+  def add_content(%Display{content: content} = display, to_add),
+    do: %{display | content: content ++ [to_add]}
+
+  defimpl Prompt.IO.Terminal do
+    def display(display) do
       _ = Prompt.raw_mode_supported?() && :shell.start_interactive({:noshell, :cooked})
 
-      # Put the cursor in the correct place
-      _ = position(text.position, text.text)
+      display =
+        if display.alt_buffer,
+          do: Display.add_content(display, Prompt.IO.ANSI.alt_screen_buffer_on()),
+          else: display
 
-      if text.mask_line && text.from == :self do
-        IO.ANSI.format([
-          :reset,
-          background_color(text),
-          text.color,
-          text.text,
-          :reset,
-          " [Press Enter to continue]"
+      display =
+        display
+        |> Display.add_content([position(display.position, display.text)])
+        |> Display.add_content([
+          display.background_color,
+          display.color,
+          display.text,
+          :reset
         ])
-        |> write()
-      else
-        [
-          :reset,
-          background_color(text),
-          text.color,
-          text.text,
-          :reset,
-          without_newline(text.trim)
-        ]
-        |> IO.ANSI.format()
-        |> IO.puts()
-      end
 
-      text
+      display =
+        (display.mask_line && display.from == :self &&
+           Display.add_content(display, [" [Press Enter to continue]"])) ||
+          maybe_with_newline(display)
+
+      display.content
+      |> IO.ANSI.format()
+      |> Prompt.IO.write()
+
+      display
     end
 
-    def evaluate(text) do
-      if text.mask_line && text.from == :self do
+    def evaluate(display) do
+      if display.mask_line && display.from == :self do
         case read(:stdio, :line) do
           :eof -> :error
           {:error, _reason} -> :error
-          _ -> Prompt.Position.mask_line(1)
+          _ -> Prompt.IO.Position.mask_line(1)
         end
       end
 
       :ok
     end
 
-    defp without_newline(true), do: ""
-    defp without_newline(false), do: "\n"
+    defp maybe_with_newline(%Display{trim: true} = display), do: display
+    defp maybe_with_newline(display), do: Display.add_content(display, "\n")
 
-    defp position(:left, _), do: write(IO.ANSI.cursor_left(10_000))
+    defp position(:left, _), do: IO.ANSI.cursor_left(10_000)
 
-    defp position(:right, content) do
+    defp position(:right, content) when is_binary(content) do
       move_left = String.length(content)
-      write(IO.ANSI.cursor_right(10_000) <> IO.ANSI.cursor_left(move_left))
+      [IO.ANSI.cursor_right(10_000), IO.ANSI.cursor_left(move_left)]
     end
 
-    defp background_color(display) do
-      case display.background_color do
-        nil -> IO.ANSI.default_background()
-        res -> String.to_atom("#{Atom.to_string(res)}_background")
-      end
+    defp position(:right, content) when is_list(content) do
+      move_left =
+        content
+        |> List.flatten()
+        |> Enum.reject(fn
+          input when is_atom(input) -> true
+          <<first::binary-size(1), _rest::binary>> when is_escape_code(first) -> true
+          _ -> false
+        end)
+        |> Enum.reduce(0, fn x, acc -> acc + String.length(x) end)
+
+      [IO.ANSI.cursor_right(10_000), IO.ANSI.cursor_left(move_left)]
     end
   end
 end