+++ lib/prompt/io/password.ex
@@ -2,70 +2,78 @@ defmodule Prompt.IO.Password do
   @moduledoc false
 
   alias __MODULE__
-  import IO, only: [write: 1, read: 2]
+  import IO, only: [read: 2]
 
   @type t :: %Password{
+          content: list(),
           text: binary(),
           color: Prompt.IO.Color.t(),
           background_color: Prompt.IO.Color.t()
         }
 
-  defstruct [:text, :color, :background_color]
+  defstruct [:content, :text, :color, :background_color]
 
   def new(text, options) do
+    background_color =
+      options
+      |> Keyword.get(:background_color)
+      |> Prompt.IO.background_color()
+
     %Password{
+      content: [:reset],
       text: text,
       color: Keyword.get(options, :color, IO.ANSI.default_color()),
-      background_color: Keyword.get(options, :background_color)
+      background_color: background_color
     }
   end
 
-  defimpl Prompt.IO do
-    def display(password) do
-      _ = Prompt.raw_mode_supported?() && :shell.start_interactive({:noshell, :cooked})
+  def add_content(%Password{content: content} = password, to_add),
+    do: %{password | content: content ++ [to_add]}
 
-      [
-        :reset,
-        background_color(password),
-        password.color,
-        password.text,
-        ": ",
-        IO.ANSI.conceal()
-      ]
+  defimpl Prompt.IO.Terminal do
+    def display(password) do
+      _ = Prompt.raw_mode_supported?() && :shell.start_interactive({:noshell, :raw})
+
+      password =
+        Password.add_content(password, [
+          password.background_color,
+          password.color,
+          password.text,
+          ": ",
+          IO.ANSI.conceal()
+        ])
+
+      password.content
       |> IO.ANSI.format()
-      |> write()
+      |> Prompt.IO.write()
 
       password
     end
 
     def evaluate(_password) do
-      case read(:stdio, :line) do
+      # case read(:stdio, :line) do
+      case (Prompt.raw_mode_supported?() && :io.get_password()) || read(:stdio, :line) do
         :eof ->
-          write(IO.ANSI.reset())
+          Prompt.IO.write(IO.ANSI.reset())
           :error
 
         {:error, _reason} ->
-          write(IO.ANSI.reset())
+          Prompt.IO.write(IO.ANSI.reset())
           :error
 
         answer when is_binary(answer) ->
-          write(IO.ANSI.reset())
+          Prompt.IO.write(IO.ANSI.reset())
           String.trim(answer)
 
         answer when is_list(answer) ->
-          write(IO.ANSI.reset())
+          Prompt.IO.write(IO.ANSI.reset())
 
           answer
           |> IO.chardata_to_string()
           |> String.trim()
       end
-    end
-
-    defp background_color(password) do
-      case password.background_color do
-        nil -> IO.ANSI.default_background()
-        res -> String.to_atom("#{Atom.to_string(res)}_background")
-      end
+    after
+      Prompt.raw_mode_supported?() && :shell.start_interactive({:noshell, :cooked})
     end
   end
 end